mcp-server/requirements-python313.txt
# Requirements for Python 3.13 compatibility
# Updated versions that support Python 3.13

# Core Framework
mcp[cli]>=1.10.1
pydantic>=2.11.7

# Data Processing & Analysis
pandas>=2.2.0
numpy>=2.1.0
pyarrow>=20.0.0

# Data Science & Machine Learning
scikit-learn>=1.7.0
scipy>=1.15.0
pyod>=2.0.5

# Data Quality & Monitoring
evidently>=0.7.9
# ydata-profiling doesn't support Python 3.13 yet, using alternative
pandas-profiling>=3.2.0
missingno>=0.5.2

# Visualization
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=5.24.0

# Development & Environment
python-dateutil>=2.9.0

# Additional dependencies for Python 3.13
typing-extensions>=4.8.0

# Master Orchestrator Dependencies
# LLM & Validation
guardrails-ai>=0.5.0
openai>=1.0.0

# Web Framework & API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0

# Messaging & Database
confluent-kafka>=2.3.0
motor>=3.3.0
pymongo>=4.6.0

# Caching & Storage
redis>=5.0.0
aioredis>=2.0.0

# Configuration & Parsing
pyyaml>=6.0.1
python-multipart>=0.0.6

# Retry & Rate Limiting
tenacity>=8.2.0
slowapi>=0.1.9

# Security & Validation
bleach>=6.1.0
validators>=0.22.0 